sdks/python/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "eks_blueprints_python_sdk"
version = "0.1.0"
authors = [
    {name = "Zachary Jacobson", email = "zjacobso@amazon.com"}
]
description = "Python Client for EKS Blueprint Multi Language Framework"
readme="README.md"
license = {text ="Apache-2.0"}
dependencies = [
    "grpcio",
    "grpcio-tools",
    "protobuf",
]

[project.urls]
"GitHub" = "https://github.com/zjaco13/multi-lang-eks-blueprints-cdk"
